fix: modal State Context -> zustand로
